+++ app/domains/content/stage1/router.py
@@ -10,7 +10,7 @@ from app.domains.content.stage1.schemas import (
 )
 from app.domains.content.stage1.service import stage1_pair_response
 from app.domains.progress.dependency.dependencies import get_learning_record_service
-from app.domains.progress.service.service import LearningRecordService
+from app.domains.progress.service.learning_record_service import LearningRecordService
 from app.infrastructure.db.mongo.mongo_client import get_mongo_client
 
 router = APIRouter(prefix="/student/learning", tags=["student-learning"])
