+++ app/domains/progress/dependency/dependencies.py
@@ -1,6 +1,6 @@
 from app.domains.classroom.repository.repository import MongoClassroomRepository
 from app.domains.progress.repository.repository import MongoLearningRecordRepository
-from app.domains.progress.service.service import LearningRecordService
+from app.domains.progress.service.learning_record_service import LearningRecordService
 from app.infrastructure.db.mongo.mongo_client import get_mongo_client
