+++ tests/agent/test_session_service.py
@@ -6,7 +6,7 @@ import pytest
 
 from app.domains.agent.models import AgentDecision, ChatMessage, ChatSession
 from app.domains.agent.repository.repository import ChatSessionRepository
-from app.domains.agent.service.service import AgentService, ChatSessionService
+from app.domains.agent.service.agent_service import AgentService, ChatSessionService
 from app.domains.progress.models import StudentWeaknessProfile, WeakConcept
