+++ tests/structure/test_domain_structure.py
@@ -13,17 +13,17 @@ def test_domain_packages_are_importable():
         "app.domains.auth.repository.repository",
         "app.domains.auth.controller.auth_router",
         "app.domains.auth.schema.schemas",
-        "app.domains.auth.service.service",
+        "app.domains.auth.service.auth_service",
         "app.domains.classroom.controller.teacher_class_router",
         "app.domains.classroom.controller.student_class_router",
         "app.domains.progress.controller.teacher_student_router",
         "app.domains.classroom.dependency.dependencies",
         "app.domains.classroom.models",
         "app.domains.classroom.repository.repository",
-        "app.domains.classroom.service.service",
+        "app.domains.classroom.service.classroom_service",
         "app.domains.classroom.schema.schemas",
         "app.domains.agent.schema.schemas",
-        "app.domains.agent.service.service",
+        "app.domains.agent.service.agent_service",
         "app.domains.instruction.controller.student_router",
         "app.domains.instruction.controller.teacher_router",
         "app.domains.instruction.service.generation",
@@ -31,7 +31,7 @@ def test_domain_packages_are_importable():
         "app.domains.instruction.models",
         "app.domains.instruction.repository.repository",
         "app.domains.instruction.schema.schemas",
-        "app.domains.instruction.service.service",
+        "app.domains.instruction.service.instruction_service",
         "app.domains.content.controller.catalog_router",
         "app.domains.content.stage1.router",
         "app.domains.content.stage1.schemas",
@@ -45,11 +45,11 @@ def test_domain_packages_are_importable():
         "app.domains.content.models",
         "app.domains.content.repository.repository",
         "app.domains.content.schema.schemas",
-        "app.domains.content.service.service",
+        "app.domains.content.service.catalog_service",
         "app.domains.progress.models",
         "app.domains.progress.repository.repository",
         "app.domains.progress.schema.schemas",
-        "app.domains.progress.service.service",
+        "app.domains.progress.service.learning_record_service",
         "app.domains.content.stage3.schemas",
         "app.domains.content.stage3.service",
         "app.domains.agent",
